feat: first-time install without user interaction
The first time install fails 2-3 times before going through, as there
are steps required before mise is actually usable with our
configurations. Mainly, the configuration file must first be trusted and
experimental features explicitly allowed.

Mise helpfully prompts the required steps to fix both of these, but does
not perform these automagically, by-itself.

To avoid having to go through the run-fix-run-again cycle on first
install, perform updating the settings and trusting the config file as
part of the start script itself. I consider this as part of the required
environment setup, thus the start script is a natural place to manage
this.