chore: emphasize doc role of the startup script
The startup script doubles as documentation for "how to set up the
development environment". The README contains some unnecessarily
overlapping descriptions of startup steps, some of which are arguably
already out of date.

To avoid the documentation getting out of date, describe the role of the
startup script as executable documentation. That is, the 'documentation'
cannot then get out of date, as it is executed on daily-basis.