validate stored tutkintotasot
Encode tutkintotasot as enums in the in-memory model and validate the DB
values to be valid tutkintotasot enumerants. Uniqueness of the array
members is not validated, so stored value can in theory contain
duplicates. In-memory model encodes tasot as a set, thus write/read
operations through the backend will still enforce uniqueness.

Using the enum directly with Spring CRUDRepository is proving difficult
as usual `@ElementCollection` and `@Enumerated` annotations are not
enough to get the serialization to figure out what it is expected to do.