add oid type facade
Implement a simple "facade" OID type to serve as the OID type in the
in-memory model. The implementation is facade over org.ietf.jgss.Oid due
to the fact that implementing the parsing is somewhat error-prone, so
using a readily available type is preferable to avoid doing unnecessary
work. However, depending on such "random" package directly to provide
the type is a nighmare for upgrading dependencies on the long run, thus
a facade helps this by keeping any possible breaking changes in a single
class.

TL;DR: I was too lazy to implement OID parsing, so I used OID class from
a readily available dependency (org.ietf.jgss). To avoid expensive
breakages from changes in external deps, I abstracted the direct
dependency behind a facade type. If the external dependency ever
changes, only the facade needs to be updated, not every single usage of
OID.