remove openapi/swagger
with only a single (internal-ish) API endpoint, there is little to no
benefit from having the OpenAPI spec and swagger UI configured. On the
contrary, there is a non-zero maintenance cost from maintaining these,
as proven by the recent "patch version with breaking changes" debugging
-situation.

So, to cut complexity and maintenance overhead, its better to ditch the
OpenAPI setup we have now. It can always be restored later on, if/when
needed.