lenient schedule parsing as "extension method"
Not quite extension method, as AFAIK there is no clean way of writing
extension methods on companion objects on Java classes (= static
extensions on Java classes). That is due to non-Kotlin classes not
having companion objects, thus there is no "static" context in Kotlin
terms for those.

Easy way out seemed to be to extend the static `Schedules` class (as it
is luckily not `final`), allowing writing the "static extension method"
in the wrapper class instead.

I would have liked to call the "extension method" `parseSchedule` to
shadow the non-lenient base implementation, but this is sadly not
possible and results in compilation failure.