avoid using oid wrapper constructor directly
The constructor is intended to be private, with instances of Oid wrapper
only created through the parsing methods, provided through the companion
object.

Relying on the "private" constructor creates a direct dependency on the
underlying org.ietf.jgss.Oid -type, which breaks the whole purpose of
the facade. The reason why the facade exists in the first place is to
hide the underlying validation/storage method of the Oids, as not to
force us to a specific implementation (and to avoid having to implement
oid parsing ourselves). Long story short, maintainability reasons.

Might be that all paths to Hell are paved with good intentions, but I'm
still somewhat confident the facade/alias/wrapper is a good fit for this
situation.