make the oid constructor actually private
To avoid anyone accidentally trying to construct instances of Oid
directly without the explicit parsing methods, make the constructor
private.

Kotlin has some funkiness with the semantics of the copy-constructor,
which requires forcing a recent enough Kotlin API version. I don't see a
reason why we would not want to use the latest, thus I just slapped in
the mvn property and called it a day.

The newer Kotlin API version allows using the @ConsistentCopyVisibility
annotation, which hides the copy-constructor alongside the canonical
constructor, when marking the said constructor as private.