add field for explicitly defining sort url params
Relying on the implicitly defined enum variant name is a bit magical,
and it is rather surprising that the lowercased name is what ends up in
the URL (or at least is expected to end up there). Furthermore, the
WebMvcConfigurer using the name for finding the correct variant when
deserializing is an additional bit of well hidden archaic magic, making
it quite easy to break the sorting, if one is not familiar with the
implementation.

So, I added a new field which explicitly defines the url param value,
and renamed all usages of the lowercaseName to use this new field
instead.

This should hopefully make things slightly less magical going forward.