Use `aqua:maven` so renovate knows how to update it
Renovate only supports certain "short names" like `maven` and only for a specific backend. The `maven` short name is supported only if using the `asdf` backend, but we're using `aqua` nowadays (as one should). This is a deficiency in upstream renovate, but we can work around it by not using a short name and specifying the backend explicitly.

Ref: https://docs.renovatebot.com/modules/manager/mise/#limitations