refact: unnecessary use of elvis operator
There are two separate "issues" on the fixed lines. These are not real
issues, but more of code-smells and code quality nitpicks:
1. empty lists being created with listOf() instead of the idiomatic
   emptyList() constructor.
2. elvis operator used where .orEmpty() extension method could have been
   used instead.

The first one, while seemingly just more characters to write the same
thing (listOf() vs. emptyList()), has actual implementation changes.
While the end result is functionally the same, the emptyList() method
returns a special singleton instance of an EmptyList, while listOf()
ends up creating an instance of an actual concrete list type. Here the
performace/memory footprint is negligible, but it is a good practice to
use the semantically correct emptyList(), where appropriate.

The second one is just as silly; the elvis operator can be replaced with
a simple .orEmpty(), which actually removes the first issue altogether,
as the empty list instantiation is then an implementation detail of the
extension method. The function call looks slightly illegal, as we have
optional chaining (?.) followed by a regular method call. However,
inspected more closely, this is completely fine, as the .orElse() is
actually implemented as an extension method of a nullable "List<T>?", so
calling it on a null instance is perfectly valid.

The end result is semantically more correct and avoids unnecessary
instantiations/allocations for creating an empty list instance.

Yes, it is the vacation season and things are quite quiet here in
Helsinki after the midsummer.