db: add fi_FI.utf8 locale
We want to sort query results with Finnish collation:

    aa
    bb
    ää

Not:

    aa
    ää
    bb

We can not change the database locale after creation, but we can create a new copy of the database via `CREATE DATABASE ... TEMPLATE ...`.

An additional wrinkle is that Postgres supports different locale providers like libc and ICU, and we can not change the locale provider of a database when copying.

Our databases in AWS use the libc locale provider, so we'll need to continue using that provider if we use this method to fix the ordering issue.

This PR adds the libc fi_FI.utf8 locale to our local development environment image so it matches our production instances.