feat: use organisation oid of OPH for logged in user
Use OID of Opetushallitus for every logged in user of this application.
It is done, because right now only users of Opetushallitus are expected to use this application.

If you are reading this commit message and that is no longer the case,
(but application is still setting organization oid of OPH for every user,)
then you should implement fetching correct organization OD for the logged in user.